+++ test/test.index.ts
@@ -42,5 +42,6 @@ describe('index', () => {
     assert(gal.IdentityPoolClient);
     assert(gal.AwsClient);
     assert(gal.BaseExternalAccountClient);
+    assert(gal.DownscopedClient);
   });
 });